requirements.txt
asgiref==3.5.2
boto3==1.24.18
botocore==1.27.19
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==38.0.1
Django==4.0.5
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.13.1
ecdsa==0.18.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
jmespath==1.0.1
Pillow==9.2.0
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2022.1
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
